Upload only difference from sftp
