+++ dags/bin/sync_sftp
@@ -3,24 +3,36 @@
 # Download all files from sftp server
 #
 # Usage> 
-# ./sync_folder.expect xxx.com 9992 agentiq-ops 1234 
+# ./sync_ftp xxx.com 9992 agentiq-ops 1234 
 #
 
 set HOST [lindex $argv 0]
 set PORT [lindex $argv 1]
 set USER [lindex $argv 2]
 set PASSWORD [lindex $argv 3]
+set MODE [lindex $argv 4]
  
 set timeout -1
  
-spawn sftp -2 -P $PORT -oHostKeyAlgorithms=+ssh-dss $USER@$HOST
+spawn sftp -2 -P $PORT -oHostKeyAlgorithms=ssh-dss -o StrictHostKeyChecking=no -o PubkeyAuthentication=no $USER@$HOST
  
 expect "$USER@$HOST's password:"
 send -- "$PASSWORD\r"
  
-expect "Connected to transfer.arvest.com.\r"
 expect "sftp>"
-send -- "get *\r"
+if { $MODE == "list" } {
+  send -- "ls -1 *.csv\r"
+} elseif { $MODE  == "fetch" } {
+  set argsCount [llength $argv]
+  set i 5
+  while {$i < $argsCount } {
+    set FILE [lindex $argv $i]
+    send -- "get $FILE\r"
+    set i [expr $i+1];
+  }
+} else {
+  send -- "pwd\r"
+}
  
 expect "sftp>"
 send -- "exit\r"
